Codesign w/ debugging entitlement for backtraces on macOS
